alerter added to delete action

+++ src/app/pages/movie-detail/movie-detail.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { MoviesService } from 'src/services/movies.service';
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-movie-detail',
   templateUrl: './movie-detail.component.html',
@@ -11,7 +12,8 @@ export class MovieDetailComponent {
   movie: any;
   constructor(
     private activatedRoute: ActivatedRoute,
-    private moviesService: MoviesService
+    private moviesService: MoviesService,
+    private router: Router
   ) {
     this.activatedRoute.paramMap.subscribe((params) => {
       this.id = params.get('id');
@@ -22,7 +24,22 @@ export class MovieDetailComponent {
     });
   }
   ngOnInit(): void {}
+  Swal = require('sweetalert2');
   delete() {
-    this.moviesService.deleteMovie(this.id).subscribe();
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#FFCA2B',
+      cancelButtonColor: '#555555',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.moviesService.deleteMovie(this.id).subscribe();
+        Swal.fire('Deleted!', 'Your file has been deleted.', 'success');
+        this.router.navigate(['/movies']);
+      }
+    });
   }
 }
